Revert "FINERACT-2181: Some tweaks in the gradle build to ensure correct task processing"
This reverts commit 11cecb0d45bef2e6f7fda3a7c3575be53732a078.
