+++ build.gradle
@@ -365,6 +365,10 @@ configure(project.fineractJavaProjects) {
         withJavadocJar()
     }
 
+    tasks.withType(ProcessResources).configureEach {
+        destinationDir = layout.buildDirectory.dir('classes/java/main').get().asFile
+    }
+
     // Add performance optimizations
     configurations.configureEach {
         resolutionStrategy {
