+++ fineract-command/dependencies.gradle
@@ -41,9 +41,6 @@ dependencies {
             'com.ibm.icu:icu4j',
             'org.yakworks:spring-icu4j',
             )
-    implementation('org.eclipse.persistence:org.eclipse.persistence.jpa') {
-        exclude group: 'org.eclipse.persistence', module: 'jakarta.persistence'
-    }
 
     compileOnly 'org.projectlombok:lombok'
     annotationProcessor 'org.projectlombok:lombok'
